--- .cpanel.yml
@@ -1,5 +1,0 @@
----
-deployment:
-  tasks:
-    - export DEPLOYPATH=/home/johnupug/public_html/website
-    - /bin/cp -R * $DEPLOYPATH
