+++ .gitignore
@@ -21,4 +21,7 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-/.firebase
+.firebase
+
+.cache
+/public
