+++ gatsby-config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  plugins: [`gatsby-plugin-sass`, `gatsby-plugin-mdx`]
+}
