+++ package.json
@@ -9,13 +9,19 @@
     "@fortawesome/free-brands-svg-icons": "^5.13.0",
     "@fortawesome/free-solid-svg-icons": "^5.13.0",
     "@fortawesome/react-fontawesome": "^0.1.9",
+    "@mdx-js/mdx": "^1.6.5",
+    "@mdx-js/react": "^1.6.5",
     "@popperjs/core": "^2.4.0",
     "bootstrap": "^4.4.1",
     "font-awesome": "^4.7.0",
+    "gatsby": "^2.23.3",
+    "gatsby-plugin-layout": "^1.3.4",
+    "gatsby-plugin-mdx": "^1.2.15",
+    "gatsby-plugin-sass": "^2.3.4",
     "husky": "^4.2.5",
     "jquery": "^3.5.0",
     "marked": "^1.0.0",
-    "node-sass": "^4.13.1",
+    "node-sass": "^4.14.1",
     "npm-why": "^1.1.9",
     "react": "^16.13.1",
     "react-bootstrap": "^1.0.0",
@@ -34,7 +40,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "develop": "gatsby develop"
   },
   "eslintConfig": {
     "extends": "react-app"
