--- src/Components/Blog/Posts/First.md
@@ -1,5 +1,0 @@
-# Hello, World!
-
-Or at least, whatever's left of it.
-
-I'll blog here, and stuff.  In the meantime, check out my [Media Page](/Media) for my 🔥 mixtape 🔥
