+++ src/Components/Blog/posts/First.mdx
@@ -1,0 +1,5 @@
+# Hello, World!
+
+Or at least, whatever's left of it.
+
+I'll blog here, and stuff.  In the meantime, check out my [Media Page for my 🔥 mixtape 🔥](/Media)
