+++ src/Components/layout/Main.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { Switch, Route } from "react-router-dom";
+
+import { Home } from "../pages/Home";
+import Personal from "../pages/Personal";
+import Professional from "../pages/Professional";
+import Media from "../pages/Media";
+
+const Main = () => {
+  return (<div>rip</div>
+
+  );
+};
+
+export default Main;
