+++ src/Components/layout/colors.scss
