+++ src/pages/index.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { Home } from "./Home";
+
+export default function Index() {
+  return (<Home />)
+}
